projects/resume-agent/requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# PDF processing
pypdf>=3.0.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# PDF generation
reportlab>=4.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
# Optional: For JavaScript-rendered pages (requires Chrome/Chromium)
# selenium>=4.0.0

# Utilities
typing-extensions>=4.8.0
